+++ src/Helper/Mpdf/Cache.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace GFPDF\Helper\Mpdf;
+
+use GFPDF_Vendor\Mpdf\Cache as MpdfCache;
+
+/**
+ * @since 6.13.0
+ */
+class Cache extends MpdfCache {
+
+	/**
+	 * Override to ensure folder permissions are being set based on the parent directory
+	 *
+	 * @param $basePath
+	 *
+	 * @return bool
+	 *
+	 * @since 6.13.0
+	 */
+	protected function createDirectory( $basePath ) {
+		if ( ! wp_mkdir_p( $basePath ) ) {
+			return false;
+		}
+
+		return true;
+	}
+}
