+++ tests/phpunit/unit-tests/test-installer.php
@@ -172,7 +172,6 @@ class Test_Installer extends WP_UnitTestCase {
 		$this->assertDirectoryExists( $gfpdf->data->template_font_location );
 		$this->assertDirectoryExists( $gfpdf->data->template_tmp_location );
 		$this->assertDirectoryExists( $gfpdf->data->mpdf_tmp_location );
-		$this->assertDirectoryExists( $gfpdf->data->mpdf_tmp_location . '/ttfontdata' );
 
 		$this->assertFileExists( $gfpdf->data->template_tmp_location . '.htaccess' );
 		$this->assertFileExists( $gfpdf->data->template_tmp_location . 'index.html' );
